Refactor: Seperate business logic through custom hook in `Header` #61
